🤖 Merge PR #52358 [Luxon] Improve DX by adding docs by @CarsonF
* [Luxon] Add docs to improve DX

* Add some tests for the couple type changes I made
