+++ types/luxon/src/luxon.d.ts
@@ -1,0 +1,9 @@
+export const VERSION: string;
+
+export * from './datetime';
+export * from './duration';
+export * from './info';
+export * from './interval';
+export * from './settings';
+export * from './zone';
+export * from './misc';
